backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
chromadb==0.5.3
sentence-transformers==3.0.1
pymupdf>=1.24.11,<2 
beautifulsoup4==4.12.3
requests==2.32.3
groq==0.9.0
python-dotenv==1.0.1
python-multipart==0.0.9
httpx==0.27.2